refactor: 메시지 서비스를 조회/커맨드로 분리 및 채팅내역 조달간 암호화/복호화 적용
