+++ src/main/java/com/sottie/sottiechat/service/MessageService.java
@@ -16,7 +16,6 @@ import org.springframework.data.mongodb.core.query.Update;
 import org.springframework.stereotype.Service;
 
 import java.time.LocalDateTime;
-import java.util.List;
 
 import static com.sottie.sottiechat.domain.SocketEvent.*;
 
@@ -25,6 +24,8 @@ import static com.sottie.sottiechat.domain.SocketEvent.*;
 @Slf4j
 public class MessageService {
     private final ChatMessageRepository chatMessageRepository;
+    private final MessageQueryService messageQueryService;
+    private final CryptoService cryptoService;
     private final RabbitTemplate rabbitTemplate;
     private final MongoTemplate mongoTemplate;
     private static final String SUBSCRIBED_EXCHANGE_NAME = "sottie.chat.exchange";
@@ -34,7 +35,7 @@ public class MessageService {
 
     public void enterChatRoom(Long roomId, MessageRequest.Enter request) {
         log.info("[채팅방 {}번] 사용자 {} 접속", roomId, request.getUserId());
-        if (isAlreadyEnteredChatRoom(roomId, request.getUserId())) {
+        if (messageQueryService.isAlreadyEnteredChatRoom(roomId, request.getUserId())) {
             return;
         }
         MessageResponse.Chat response = MessageResponse.Chat.from(request);
@@ -71,13 +72,12 @@ public class MessageService {
     public void updateLastReadStatus(Long roomId, MessageRequest.LastRead lastRead) {
         // 채팅방 접속하는 경우, 최신 채팅 조회
         if (lastRead.getMessageId() == null) {
-            String latestChatId = getLatestChatId(roomId, lastRead.getUserId());
+            String latestChatId = messageQueryService.getLatestChatId(roomId, lastRead.getUserId());
             if (latestChatId == null)
                 return;
             lastRead.setMessageId(latestChatId);
         }
-        ChatMessage message = chatMessageRepository.findById(lastRead.getMessageId()).orElseThrow(() -> new IllegalArgumentException("존재하지 않는 메시지입니다."));
-
+        ChatMessage message = messageQueryService.getMessage(lastRead.getMessageId());
         // 일반 채팅이 아닌 경우 읽음 처리 무시
         if (message.getChatType() != ChatType.CHAT) {
             log.warn("[메시지 읽음 처리]: ChatType이 CHAT이 아님");
@@ -102,16 +102,13 @@ public class MessageService {
         }
     }
 
-    public List<LastReadStatus> getReadStatusByChatRoom(Long roomId) {
-        return mongoTemplate.find(new Query(Criteria.where("chatRoomId").is(roomId)), LastReadStatus.class);
-    }
-
     private void sendMessageToSubs(Long roomId, MessageResponse.Chat response, String routingKey, SocketEvent event) {
         Status status = Status.SUCCESS;
         // 저장하고 저장된 메시지 ID까지 함께 구독자들에게 전파
         ChatMessage saved = null;
         try {
-            saved = chatMessageRepository.save(ChatMessage.from(roomId, response, status));
+            String encoded = cryptoService.encodeAES(response.getContents());
+            saved = chatMessageRepository.save(ChatMessage.from(roomId, response, status, encoded));
             response.updateMessageId(saved.getId());
             rabbitTemplate.convertAndSend(SUBSCRIBED_EXCHANGE_NAME, routingKey + roomId, new CommonResponse<>(event, response));
         } catch (AmqpException e) {
@@ -122,38 +119,6 @@ public class MessageService {
             }
             log.error("[{}] AMQP Protocol Exception (publish failure): {}", routingKey, e.getMessage());
         }
-    }
-
-    public List<MessageResponse.Chat> getChatMessagesByPaging(Long roomId, String cursor, int size) {
-        if (size < 1)
-            throw new IllegalArgumentException("페이지 개수가 유효하지 않습니다.");
-
-        // TODO: 일별로 response 정렬하기
-        if (cursor == null) // 최초 페이지
-            return buildMessages(chatMessageRepository.findLatestChat(roomId, size));
-
-        return buildMessages(chatMessageRepository.findChatByPaging(roomId, cursor, size));
-    }
-
-    private List<MessageResponse.Chat> buildMessages(List<ChatMessage> chatMessages) {
-        return chatMessages.stream().map((c) -> MessageResponse.Chat.builder()
-                .messageId(c.getId())
-                .contents(c.getContents())
-                .userId(c.getUserId())
-                .timestamp(c.getTimestamp().toString())
-                .messageType(c.getMessageType())
-                .chatType(c.getChatType())
-                .status(c.getStatus()).build()).toList();
-    }
-
-    private boolean isAlreadyEnteredChatRoom(Long roomId, Long senderId) {
-        return !chatMessageRepository.findByChatRoomIdAndSenderIdWithEntrance(roomId, senderId).isEmpty();
-    }
-
-    private String getLatestChatId(Long roomId, Long userId) {
-        return chatMessageRepository.findLatestChatOthers(roomId, userId)
-                .map(ChatMessage::getId)
-                .orElse(null);
     }
 
     private boolean updateLastRead(Long roomId, String messageId, Long userId) {
